==> notebooks/1_create_user_interests.ipynb <==
import numpy as np
import pandas as pd
import random
import scipy

from scipy import stats

data = pd.read_csv(r'/Users/harman/Desktop/EDA_NLP/Recommendation System/data/0_news_articles.csv')
print(data.shape)
data.head()

total_length = 2250 # dataset has 2250 rows
data = data.drop(["Category"], axis = 1)
data.head()

# UserID

def generator():
    session = stats.geom.rvs(size = 2250, p = 0.3) # dataset has 2250 rows
    return session

user_session = generator()

user_id = []
id = 1

for u_id in range(total_length):
    user_id.append(id)
    if u_id > 10:
        id += 1
        
user_id = np.array(user_id).flatten()
len(user_id)

data['UserId'] = user_id
print(data.shape)

session_id = []
i = 1

for s_id in range(total_length):
    session_id.append(i)
    if s_id > 15:
        i += 1

session_id = np.array(session_id).flatten()
len(session_id)

data['SessionId'] = session_id
print(data.shape)

# Article Rank

article_rank = []
i = 1
for ar in range(total_length):
    article_rank.append(i)
    i += 1
    
    if i>10:
        i = 1
len(article_rank)

data['Article Rank'] = article_rank

# Click status

click_session = []
for c in range(total_length):
    click_session.append(random.choice([True, False]))
    
len(click_session)

data['Click'] = click_session

# Time taken

time_session = []
for t in range(total_length):
    time_session.append(np.random.randint(10, 100))
    
# if click is false set time to 0

for i in range(total_length):
    if click_session[i] is True:
        time_session[i] = time_session[i]
    else:
        time_session[i] = 0

len(time_session)

data['Time Spent (seconds)'] = time_session

data.head()

data.tail()

for i in range(total_length):
    if time_session[i] == 0:
        data = data.drop(i)
        total_length -= 1
print(total_length)

data.head()

data.tail()

data.shape

data.to_csv(r'data/2_user_interest.csv')

## For makinng Recommendations based on Ratings, we need to generate Ratings

Method followed:
    
    1. 0 rating for least interesting article
    
    2. 5 rating for most interesting article

ratings = []
for r in range(total_length):
    ratings.append(np.random.randint(1, 6))
        
len(ratings)

### Adding Rating Dataframe

data['Ratings'] = ratings
data.head()

ratings_data = data.drop(columns = ['Description', 'Date', 'URL', 'Click'])
ratings_data.head()

ratings_data.to_csv('/data/3_user_rated_articles.csv')
