==> notebooks/5_collaborative_&_hybrid.ipynb <==
import matplotlib.pyplot as plt
import numpy as np
import pandas as pd
import string
import re

import gensim
from gensim import corpora, models, similarities

import nltk
from nltk.corpus import stopwords

from lightfm import LightFM

import scipy
from scipy.spatial.distance import cdist
from scipy.sparse import csr_matrix
from scipy.sparse.linalg import svds
from scipy.linalg import svd

from sklearn.cluster import KMeans
from sklearn.feature_extraction.text import TfidfVectorizer
from sklearn.metrics import silhouette_score, silhouette_samples
from sklearn.metrics.pairwise import cosine_similarity, linear_kernel
from sklearn.model_selection import train_test_split

# Reading data sets

1. rated_articles - Contains user_interests with ratings
2. news_articles - Contains raw articles without user data amalgamation

## news_articles

data = pd.read_csv(r'/Users/harman/Desktop/EDA_NLP/Recommendation System/data/0_news_articles.csv')
data.head()

## Collaborative Filtering

**Need:** Ratings Matrix so I generated user profile with ratings

rating = pd.read_csv(r'data/3_user_rated_articles.csv')
print(rating.shape)
rating.drop(columns= rating.columns[0], 
        axis=1, 
        inplace=True)
rating.head()

rating.tail()

n_users = int(rating.UserId.nunique())
n_article = int(rating.Article_id.nunique())
print("Number of users: " , n_users)
print("Number of articles: ", n_article)

user_pivot = rating.pivot_table(index = 'UserId', columns = 'Article_id', values = 'Ratings')
user_pivot.head()

user_pivot.shape

user_pivot = user_pivot.fillna(0)
user_pivot_matrix = user_pivot.values
user_pivot_matrix

user_item_pivot_sparse = csr_matrix(user_pivot)

n_factors = 150
U, sigma, Vt = svds(user_pivot_matrix, k = n_factors)

sigma = np.diag(sigma)
sigma.shape

all_user_ratings = np.dot(np.dot(U, sigma), Vt) 
all_user_ratings_norm = (all_user_ratings - all_user_ratings.min()) / (all_user_ratings.max() - all_user_ratings.min())
all_user_ratings_norm

cf_preds_df = pd.DataFrame(all_user_ratings_norm, columns = user_pivot.columns).transpose()

class Collaborative:
    
    name = "Collaborative Filter"
    
    def __init__(self, predictions, items = None):
        self.predictions = predictions
        self.items = items
        
    def get_model_name(self):
        return self.name
            
    def recommend_items(self, user_id, items_ignore = [], topn = 10, verbose = False):
        sorted_preds = self.predictions[user_id].sort_values(ascending = False).reset_index()

        recommendations = sorted_preds[~sorted_preds['Article_id'].isin(items_ignore)].head(topn)

        if verbose:
            if self.items is None:
                raise Exception('"items" required in verbose mode')

            recommendations = recommendations.merge(self.items, how = 'left', left_on = 'Article_id', 
                                                    right_on = 'Article_id')[['Article_id', 'Title']]


        return recommendations

model = Collaborative(cf_preds_df, data)

model.recommend_items(user_id = 224, verbose = True)

## Hybrid Recommendor System

**Using:** LightRF

user_pivot.head()

u_id = list(user_pivot.index)
u_dict = {}
counter = 0

for i in u_id:
    u_dict[i] = counter
    counter += 1

len(u_dict)

# convert to csr matrix
u_interaction_csr = csr_matrix(user_pivot.values)
u_interaction_csr

item_dict ={}
df = rating[['Article_id', 'Title']].sort_values('Article_id').reset_index()

for i in range(df.shape[0]):
    item_dict[(df.loc[i,'Article_id'])] = df.loc[i,'Title']

len(item_dict)

### LightFM not working, so hopes low

model = LightFM(loss = 'warp', random_state = 2016, learning_rate = 0.90, no_components = 150, user_alpha = 0.000005)
model = model.fit(u_interaction_csr, epochs = 100, num_threads = 16, verbose = False)

title = list(data['Title'])
title[:10]

total = data.isnull().sum().sort_values(ascending = False)
percent = (data.isnull().sum()/data.isnull().count()).sort_values(ascending = False)
missing_data = pd.concat([total, percent], axis = 1, keys = ['Total', 'Percent'])
missing_data.head()
